WIP refactoring js modules to better support javascript views
