+++ frontend/web/js/core/actions/spinner-actions.ts
