+++ frontend/web/js/core/types/api/groups.ts
@@ -1,0 +1,14 @@
+import { ColumnDefinition } from "../table";
+import { PaginatedResponse } from "../pagination";
+
+export interface GroupInfo {
+  group_id: string;
+  display_name: string;
+  description: string;
+  sample_count: number;
+  table_columns: ColumnDefinition[];
+  created_at: string;
+  modified_at: string;
+}
+
+export type ApiGroupInfoResponse = PaginatedResponse<GroupInfo>;
