+++ frontend/web/js/core/types/api/membership.ts
@@ -1,0 +1,6 @@
+export type MembershipEdge = {
+  sample_id: string;
+  group_id: string;
+};
+
+export type MembershipEdges = MembershipEdge[];
