+++ frontend/web/js/core/types/api/users.ts
@@ -1,0 +1,10 @@
+export interface ApiUserInfo {
+  id: string;
+  username: string;
+  first_name: string;
+  last_name: string;
+  email: string;
+  disabled: boolean;
+  roles: string[];
+  authentication_method: string;
+}
