+++ frontend/web/js/core/types/callbacks.ts
@@ -1,0 +1,2 @@
+export type CallbackFunc = (ids: string[]) => void;
+export type TblStateCallbackFunc = (selectedRows: string[]) => void;
