+++ frontend/web/js/core/types/enums.ts
