+++ frontend/web/js/core/types/pagination.ts
@@ -1,0 +1,5 @@
+export interface PaginatedResponse<T> {
+  data: T[];
+  records_total: number;
+  records_filtered: number;
+}
