+++ frontend/web/js/core/types/table.ts
@@ -1,0 +1,11 @@
+export interface ColumnDefinition {
+  id: string;
+  label: string;
+  path: string;
+  type: string;
+  hidden: boolean;
+  sortable: boolean;
+  filterable: boolean;
+  filter_type: string;
+  filter_param: string;
+}
