+++ frontend/web/js/views/group-editor/index.ts
@@ -1,0 +1,5 @@
+import { initGroupEditor } from "./init";
+
+document.addEventListener("DOMContentLoaded", () => {
+  initGroupEditor();
+});
